pyproject.toml
[project]
name = "my-auction"
version = "0.1.0"
description = "Korean real estate auction analysis AI agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "streamlit>=1.38.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
